Moving code from catalogd/internal to internal/catalogd
This partially fixes #1707. The intent is to consolidate internal code from operator-controller and catalogd within internal.
This PR does not move catalogd/internal/controllers and
catalogd/internal/webhook as it need explict changes in both op-con and
catalogd Makefile. Which we will do in a followup PR.

    catalogd/ - code specific to catalogd
    internal/ - holds internal code of catalogd and operator-controller.
    internal/catalogd - holds unexported code specific to catalogd
    internal/operator-controller - holds unexported code specific to operator-comntroller
    internal/shared - shared code between catalogd and operator-controller which can not be exported outside.
