+++ catalogd/cmd/catalogd/main.go
@@ -57,13 +57,13 @@ import (
 
 	catalogdv1 "github.com/operator-framework/operator-controller/catalogd/api/v1"
 	corecontrollers "github.com/operator-framework/operator-controller/catalogd/internal/controllers/core"
-	"github.com/operator-framework/operator-controller/catalogd/internal/features"
-	"github.com/operator-framework/operator-controller/catalogd/internal/garbagecollection"
-	catalogdmetrics "github.com/operator-framework/operator-controller/catalogd/internal/metrics"
-	"github.com/operator-framework/operator-controller/catalogd/internal/serverutil"
-	"github.com/operator-framework/operator-controller/catalogd/internal/source"
-	"github.com/operator-framework/operator-controller/catalogd/internal/storage"
 	"github.com/operator-framework/operator-controller/catalogd/internal/webhook"
+	"github.com/operator-framework/operator-controller/internal/catalogd/features"
+	"github.com/operator-framework/operator-controller/internal/catalogd/garbagecollection"
+	catalogdmetrics "github.com/operator-framework/operator-controller/internal/catalogd/metrics"
+	"github.com/operator-framework/operator-controller/internal/catalogd/serverutil"
+	"github.com/operator-framework/operator-controller/internal/catalogd/source"
+	"github.com/operator-framework/operator-controller/internal/catalogd/storage"
 	fsutil "github.com/operator-framework/operator-controller/internal/shared/util/fs"
 	"github.com/operator-framework/operator-controller/internal/shared/version"
 )
