+++ internal/catalogd/garbagecollection/garbage_collector.go
