+++ internal/catalogd/source/containers_image.go
