+++ internal/catalogd/storage/http_preconditions_check.go
