+++ internal/catalogd/storage/index.go
