+++ internal/catalogd/storage/index_test.go
